+++ src/pager/pager-dropdown.ts
@@ -1,6 +1,16 @@
 import { createElement, remove } from '@syncfusion/ej2-base';
 import { Pager } from './pager';
-import { DropDownList } from '@syncfusion/ej2-dropdowns';
+import { DropDownList, ChangeEventArgs } from '@syncfusion/ej2-dropdowns';
+
+/**
+ * IPager interface
+ * @hidden
+ */
+export interface IPager {
+    newProp: {
+        value: number | string | boolean
+    };
+}
 /**
  * `PagerDropDown` module handles selected pageSize from DropDownList. 
  */
@@ -44,34 +54,66 @@ export class PagerDropDown {
         this.pagerDropDownDiv.appendChild(dropDownDiv);
         this.pagerDropDownDiv.appendChild(defaultTextDiv);
         this.pagerModule.element.appendChild(this.pagerDropDownDiv);
-        let pageSizesModule: boolean | number[] = this.pagerModule.pageSizes;
-        let pageSizesArray: number[] = ((<number[]>pageSizesModule).length ? pageSizesModule : [5, 10, 12, 20]) as number[];
-        let defaultValue: Number = (pageSizesArray).indexOf(this.pagerModule.pageSize) > -1 ? this.pagerModule.pageSize : pageSizesArray[0];
+        let pageSizesModule: boolean | (number | string)[] = this.pagerModule.pageSizes;
+        let pageSizesArray: string[] = ((<string[]>pageSizesModule).length ? this.convertValue(pageSizesModule as string[]) :
+            ['5', '10', '12', '20', 'All']) as string[];
+        let defaultValue: Number | String = (pageSizesArray).indexOf(this.pagerModule.pageSize.toString()) > -1 ? this.pagerModule.pageSize
+            : (pageSizesArray[0] === 'All' ? this.pagerModule.totalRecordsCount : parseInt(pageSizesArray[0], 10));
         this.dropDownListObject = new DropDownList({
             dataSource: pageSizesArray,
-            value: defaultValue as number,
+            value: defaultValue.toString() as string,
             change: this.onChange.bind(this)
         });
         this.dropDownListObject.appendTo(input);
         pagerObj.pageSize = defaultValue as number;
         pagerObj.dataBind();
         pagerObj.trigger('dropDownChanged', { pageSize: defaultValue });
+        this.addEventListener();
     }
     /**
      * For internal use only - Get the pagesize. 
      * @private
      * @hidden
      */
-    private onChange(e: Event): void {
-        this.pagerModule.pageSize = this.dropDownListObject.value as number;
+    private onChange(e: ChangeEventArgs): void {
+        if (this.dropDownListObject.value === 'All') {
+            this.pagerModule.pageSize = this.pagerModule.totalRecordsCount;
+            this.pagerCons.innerHTML = this.pagerModule.getLocalizedLabel('pagerAllDropDown');
+            e.value = this.pagerModule.pageSize;
+        } else {
+            this.pagerModule.pageSize = parseInt(this.dropDownListObject.value as string, 10);
+            if (this.pagerCons.innerHTML !== this.pagerModule.getLocalizedLabel('pagerDropDown')) {
+                this.pagerCons.innerHTML = this.pagerModule.getLocalizedLabel('pagerDropDown');
+            }
+        }
         this.pagerModule.dataBind();
         this.pagerModule.trigger('dropDownChanged', { pageSize: this.dropDownListObject.value });
+    }
+
+    private beforeValueChange(prop: IPager): void {
+        if (typeof prop.newProp.value === 'number') {
+            let val: string = prop.newProp.value.toString();
+            prop.newProp.value = val;
+        }
+    }
+    private convertValue(pageSizeValue: (number | string)[]): (number | string)[] {
+        let item: (number | string)[] = pageSizeValue;
+        for (let i: number = 0; i < item.length; i++) {
+            item[i] = typeof item[i] === 'number' ? item[i].toString() : item[i];
+        }
+        return item as string[];
     }
 
     public setDropDownValue(prop: string, value: string | number | Object | boolean): void {
         if (this.dropDownListObject) {
             this.dropDownListObject[prop] = value;
         }
+    }
+    public addEventListener(): void {
+        this.dropDownListObject.on('beforeValueChange', this.beforeValueChange, this);
+    }
+    public removeEventListener(): void {
+        this.dropDownListObject.off('beforeValueChange', this.beforeValueChange);
     }
 
     /**
@@ -82,6 +124,7 @@ export class PagerDropDown {
      */
     public destroy(args?: { requestType: string }): void {
         if (this.dropDownListObject && !this.dropDownListObject.isDestroyed) {
+            this.removeEventListener();
             this.dropDownListObject.destroy();
             remove(this.pagerDropDownDiv);
         }